Early draft for conjugations
This constitutes an early draft for the conjugations section.
Better verb examples are needed in a few places, and the third-io
needs to be filled out for where it differs from the third.

Some tweaks done to verbchart will need to be done to the other
charts in res/macro/langtab.tex; mostly these are to do with
spacing and playing nicely with surrounding paras.

Preface is, for now, in point form, and does not reflect the order
which it will take once it is turned into paragraphs.