Add imperfect tense conjugations
The imperfect tense is now enumerated for all verb conjugations.
A greater degree of variety is required for the choices of example
words.  By this I mean that the third-io now has examples for all
drafted tenses, voices, and moods, but for the same verb in each
example.  Later on these examples will need to be looked at again
and replaced with different example verbs for each chart.