Add rudimentary linux cheat sheet example
Now present is an example which is a basic linux cheat sheet for
using Bash, Git, SSH, and Python.  It's not meant to be exhaustive
but something as a reference for a beginner, and since it's just
an example, it's going to stay small.

Also included in the commit are accompanying resource files,
such as a TiKZ command to resemble a keystroke font, and a
command much like the \dictentry command in res/macro/langtab for
a cheat sheet entry.  Also to come is a cheat sheet for vim.