Start transition from headers/macros to classes/packages
The .tex files have become too unwieldy to be .tex files.
So beginneth the change from headers to classes, and macros
to packages.  The first victim of this change is the dictionary
format, which will henceforth be called linguistics.cls.
Might change this to dictionary.cls later, as there is already
a linguistics.sty which will be used by the assignment class.