Convert coverletter.sty to .dtx
The coverletter macro package has had its format changed
to follow the LaTeX doc format, with code and documentation
in the same .dtx file.  This one file will generate the
requisite .sty, .pdf. README.txt, and .ins files needed
for package distribution and documentation.

From here, this file will be refined somewhat and adapted
into a template to facilitate conversion of the remaining
packages and classes to dtx format.