Add hypdoc to dictionary for quick reference
Got tired of scrolling through the document when I want to look at
a section (searching for words gets messed up with macrons in the
mix).  Now at least when I open the document in apvlv I get a nice
set of links with which to access the relevant sections.