Convert linguistics sty to dtx
The linguistics macro file is now in dtx format and has been tested
on the dictionary example.  Refchart has been removed as it is no
longer used anywhere, and the code is slightly reorganized and
pruned.  Still needs a means for keeping the heading and body
tables together.

Also included in the commit are corresponding changes to the dtx
template; as refinements are made to individual dtx files, the
template should reflect these refinements.  README.md has been
updated to reflect the changes in the commit.