Convert resume sty to dtx
The resume macro package for use with the resume class is in
dtx format with in-line documentation.  Some macro names may
need to be changed as they are not particularly descriptive,
but they'll work for now without breaking my current resume.

Added to the resume macro package is a double action header
which does not act as a section heading, which was previously
missing (you could have a double action heading, but it had
to be a section heading as well).