Convert ceystroke sty to dtx
The keystroke text decorator macro package is now available
in dtx format, with in-line documentation.  Also added to
the package are a couple of dimensions which allow for some
limited flexibility in the rendering of the decoration.

Minor changes to the template reflect the ongoing evolution
of the dtx format template, and the README.md file now
reflects the current state of dtx conversion, and has a
little tip to help (some) users with installation; this
will likely need adjustment in the future.