Add conjunctions section to dictionary
There is now a short section for conjunction words.  The
addition of this section gives me pause as to whether there
aren't too many dictionary sections to this unwieldy thing.
I will have to ponder the merits of having everything
mashed into one section (or two, given there may also be
an English to Latin section).