Reorder dictionary sections
As the dictionary gets bigger, it occurs to me that
separating verb entries from noun entries necessitates a
more logical structure, with pertinent information about
a class of words (i.e., nouns) preceding said class.

Not to say I won't end up trashing this organizational
structure down the line, but for now it makes a novel
reference approach, and turning it into a traditional
dictionary would make it somewhat superfluous.