Reorganize conjugations section into voices
This is a big commit which seeks to add the perfect-tense
passive voice conjugations in a logical place.  It makes sense
to precede the passive voice (or at the very least the perfect
passive) with at least some discussion on participles.  Both
sections are now somewhat dirty and will require attention
(mitigate repetitions, more elucidation needed, et cetera).

Still missing is the imperfect passive subjunctive for all
conjugations.