Rename linguistics to dictionarium
Both the linguistics class and its corresponding macro
package have been renamed to dictionarium.  This is both
an acknowledgement that the class is unlikely to encompass
a document toolkit which would be useful for linguists
(unlikely but not impossible) and to reflect the parallel
development with my Latin reference project.